pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "us-2024-political-polarization"
version = "0.1.0"
description = "End-to-end reproducible analysis pipeline for US 2024 political polarization using CCES survey data."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "Your Name" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "matplotlib>=3.8",
  "scipy>=1.12",
  "pyarrow>=15.0",
  "joblib>=1.3",
  "streamlit>=1.32",
  "plotly>=5.20",
]

[project.optional-dependencies]
umap = ["umap-learn>=0.5.6"]
hdbscan = ["hdbscan>=0.8.33"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ipykernel>=6.29"]

[project.scripts]
polarization = "polarization.__main__:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
